Se agrego llave foranea a tabla insects, por tanto la base de datos está en la versión 3
