+++ app/build.gradle
@@ -21,7 +21,6 @@ android {
         targetCompatibility = '1.8'
     }
 }
-
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'androidx.appcompat:appcompat:1.0.0-beta01'
